+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/Currencies/Commands/Create/Models.cs
@@ -5,7 +5,7 @@ public class RequestModel : IRequestModel
 	public string Code { get; set; }
 	public string Symbol { get; set; }
 	public string Name { get; set; }
-	public decimal ExchangeRateToUSD { get; set; }
+	public double ExchangeRateToUSD { get; set; }
 	public bool IsActive { get; set; } = true;
 }
 
@@ -15,7 +15,7 @@ public class ResponseModel : IResponseModel
 	public string Code { get; set; }
 	public string Symbol { get; set; }
 	public string Name { get; set; }
-	public decimal ExchangeRateToUSD { get; set; }
+	public double ExchangeRateToUSD { get; set; }
 	public DateTime LastUpdated { get; set; }
 	public bool IsActive { get; set; }
 	public DateTime CreatedAt { get; set; }
