+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/Currencies/Commands/Update/Models.cs
@@ -6,7 +6,7 @@ public class RequestModel : IRequestModel
 	public string Code { get; set; }
 	public string Symbol { get; set; }
 	public string Name { get; set; }
-	public decimal ExchangeRateToUSD { get; set; }
+	public double ExchangeRateToUSD { get; set; }
 	public bool IsActive { get; set; }
 }
 
@@ -16,7 +16,7 @@ public class ResponseModel : IResponseModel
 	public string Code { get; set; }
 	public string Symbol { get; set; }
 	public string Name { get; set; }
-	public decimal ExchangeRateToUSD { get; set; }
+	public double ExchangeRateToUSD { get; set; }
 	public DateTime LastUpdated { get; set; }
 	public bool IsActive { get; set; }
 	public DateTime? UpdatedAt { get; set; }
