+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/Currencies/Queries/All/Models.cs
@@ -13,7 +13,7 @@ public class ResponseModel : IResponseModel<Array>
 	public string Code { get; set; }
 	public string Symbol { get; set; }
 	public string Name { get; set; }
-	public decimal ExchangeRateToUSD { get; set; }
+	public double ExchangeRateToUSD { get; set; }
 	public DateTime LastUpdated { get; set; }
 	public bool IsActive { get; set; }
 	public DateTime CreatedAt { get; set; }
