+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/Offers/Commands/Create/Models.cs
@@ -5,7 +5,7 @@ public class RequestModel : IRequestModel
 	public Guid ProductId { get; set; }
 	public Guid BuyerUserId { get; set; }
 	public Guid SellerUserId { get; set; }
-	public decimal OfferedPrice { get; set; }
+	public double OfferedPrice { get; set; }
 	public string Currency { get; set; }
 	public int Quantity { get; set; }
 	public string Message { get; set; }
@@ -20,7 +20,7 @@ public class ResponseModel : IResponseModel
 	public Guid ProductId { get; set; }
 	public Guid BuyerUserId { get; set; }
 	public Guid SellerUserId { get; set; }
-	public decimal OfferedPrice { get; set; }
+	public double OfferedPrice { get; set; }
 	public string Currency { get; set; }
 	public int Quantity { get; set; }
 	public string Message { get; set; }
