+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/ProductPrices/Commands/Update/Models.cs
@@ -5,8 +5,8 @@ public class RequestModel : IRequestModel
 	public Guid Id { get; set; }
 	public Guid ProductId { get; set; }
 	public string CurrencyCode { get; set; }
-	public decimal Price { get; set; }
-	public decimal? DiscountedPrice { get; set; }
+	public double Price { get; set; }
+	public double? DiscountedPrice { get; set; }
 	public string CountryCodes { get; set; }
 	public bool IsActive { get; set; }
 	public DateTime? ValidFrom { get; set; }
@@ -19,8 +19,8 @@ public class ResponseModel : IResponseModel
 	public Guid Id { get; set; }
 	public Guid ProductId { get; set; }
 	public string CurrencyCode { get; set; }
-	public decimal Price { get; set; }
-	public decimal? DiscountedPrice { get; set; }
+	public double Price { get; set; }
+	public double? DiscountedPrice { get; set; }
 	public string CountryCodes { get; set; }
 	public bool IsActive { get; set; }
 	public DateTime? ValidFrom { get; set; }
