+++ BusinessModules/LivestockTrading/LivestockTrading.Application/RequestHandlers/SeedInfos/Commands/Create/Models.cs
@@ -9,7 +9,7 @@ public class RequestModel : IRequestModel
 	public string CommonNames { get; set; }
 	public string SeedSize { get; set; }
 	public string SeedColor { get; set; }
-	public decimal? GerminationRate { get; set; }
+	public double? GerminationRate { get; set; }
 	public int? GerminationDays { get; set; }
 	public string ClimateZones { get; set; }
 	public string SoilType { get; set; }
@@ -53,7 +53,7 @@ public class ResponseModel : IResponseModel
 	public string CommonNames { get; set; }
 	public string SeedSize { get; set; }
 	public string SeedColor { get; set; }
-	public decimal? GerminationRate { get; set; }
+	public double? GerminationRate { get; set; }
 	public int? GerminationDays { get; set; }
 	public string ClimateZones { get; set; }
 	public string SoilType { get; set; }
